se-linux: Add pixelstats access to read sysf_h2omg*
Grant read access to the files and directories of h2omg to
pixelstatus_vendor.  Because of the pixelstats is common across multiple
product products and revisions released previously and h2omg is a new
feature adding any reference to h2omg from pixelstats would require
updates to released products just to build.

The existing pixelstats_vendor.te is updated with a comment referencing
this change to provide some traceability for the future.

Bug: 390690180
Change-Id: I4d65dbb81ca6954a0ee11f0e3ddd69ee2fd8f2e6
Signed-off-by: Jim Wylder <jwylder@google.com>